Append file extension to URL.
This is purely to tell the browser how to handle the file.
The problem with Content-Disposition is it only told the browser what
file extension it had, not how to handle it.
The daemon actually strips out the file extension from the request URL
as to not get in the way of database ID lookups.
Files can still be requested by their ID alone.